Initial Commit
This is the first playable and bug-free version of the game. I was headed in the right direction in regards to OOP design, but ran out of time to be able to take a proper approach (that is why you will see many hard coded values and giant if else block statements).  Most of the subsequent changes to this repo will more so involve design changes rather than changes in logic.